+++ test/encodings/onehot/test_oneHotEncoder.py
@@ -188,100 +188,6 @@ class TestOneHotEncoder(unittest.TestCase):
         shutil.rmtree(path)
 
 
-    def construct_test_sequencedataset(self, path):
-        sequences = [ReceptorSequence(amino_acid_sequence="AAATTT", identifier="1", metadata=SequenceMetadata(custom_params={"l1": 1})),
-                     ReceptorSequence(amino_acid_sequence="ATATAT", identifier="2", metadata=SequenceMetadata(custom_params={"l1": 2}))]
-
-        PathBuilder.build(path)
-        filename = "{}sequences.pkl".format(path)
-        with open(filename, "wb") as file:
-            pickle.dump(sequences, file)
-
-        lc = LabelConfiguration()
-        lc.add_label("l1", [1, 2])
-
-        return SequenceDataset(params={"l1": [1, 2]}, filenames=[filename], identifier="d1")
-
-
-
-    def construct_test_receptordataset(self, path):
-        receptors = [TCABReceptor(alpha=ReceptorSequence(amino_acid_sequence="AAATTT", identifier="1a"),
-                                  beta=ReceptorSequence(amino_acid_sequence="ATATAT", identifier="1b"),
-                                  metadata={"l1": 1},
-                                  identifier="1"),
-                     TCABReceptor(alpha=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2a"),
-                                  beta=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2b"),
-                                  metadata={"l1": 2},
-                                  identifier="2"),
-                     TCABReceptor(alpha=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2a"),
-                                  beta=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2b"),
-                                  metadata={"l1": 2},
-                                  identifier="2")]
-
-        return ReceptorDataset.build(receptors, 10, "{}receptors.pkl".format(path))
-
-    def test_sequence_flattened(self):
-        path = EnvironmentSettings.root_path + "test/tmp/onehot_seq_flat/"
-
-        PathBuilder.build(path)
-
-        dataset = self.construct_test_sequencedataset(path)
-
-        encoder = OneHotEncoder.build_object(dataset, **{"use_positional_info": False, "distance_to_seq_middle": None, "flatten": True})
-
-        encoded_data = encoder.encode(dataset, EncoderParams(
-            result_path=path,
-            label_config=LabelConfiguration([Label(name="l1", values=[1, 0], positive_class="1")]),
-            pool_size=1,
-            learn_model=True,
-            model={},
-            filename="dataset.pkl"
-        ))
-
-        self.assertTrue(isinstance(encoded_data, SequenceDataset))
-
-        onehot_a = [1.0] + [0.0] * 19
-        onehot_t = [0.0] * 16 + [1.0] + [0] * 3
-
-        self.assertListEqual(list(encoded_data.encoded_data.examples[0]), onehot_a+onehot_a+onehot_a+onehot_t+onehot_t+onehot_t)
-        self.assertListEqual(list(encoded_data.encoded_data.examples[1]), onehot_a+onehot_t+onehot_a+onehot_t+onehot_a+onehot_t)
-
-        self.assertListEqual(list(encoded_data.encoded_data.feature_names), [f"{pos}_{char}" for pos in range(6) for char in EnvironmentSettings.get_sequence_alphabet()])
-        shutil.rmtree(path)
-
-    def test_receptor_flattened(self):
-        path = EnvironmentSettings.root_path + "test/tmp/onehot_recep_flat/"
-
-        PathBuilder.build(path)
-
-        dataset = self.construct_test_receptordataset(path)
-
-        encoder = OneHotEncoder.build_object(dataset, **{"use_positional_info": False, "distance_to_seq_middle": None,
-                                                         "flatten": True})
-
-        encoded_data = encoder.encode(dataset, EncoderParams(
-            result_path=path,
-            label_config=LabelConfiguration([Label(name="l1", values=[1, 0], positive_class="1")]),
-            pool_size=1,
-            learn_model=True,
-            model={},
-            filename="dataset.pkl"
-        ))
-
-        self.assertTrue(isinstance(encoded_data, ReceptorDataset))
-
-        onehot_a = [1.0] + [0.0] * 19
-        onehot_t = [0.0] * 16 + [1.0] + [0] * 3
-
-        self.assertListEqual(list(encoded_data.encoded_data.examples[0]), onehot_a+onehot_a+onehot_a+onehot_t+onehot_t+onehot_t+onehot_a+onehot_t+onehot_a+onehot_t+onehot_a+onehot_t)
-        self.assertListEqual(list(encoded_data.encoded_data.examples[1]), onehot_a*12)
-        self.assertListEqual(list(encoded_data.encoded_data.examples[2]), onehot_a*12)
-
-        self.assertListEqual(list(encoded_data.encoded_data.feature_names), [f"{chain}_{pos}_{char}" for chain in ("alpha","beta") for pos in range(6) for char in EnvironmentSettings.get_sequence_alphabet()])
-
-        shutil.rmtree(path)
-
-
     def test_repertoire_flattened(self):
         path = EnvironmentSettings.root_path + "test/tmp/onehot_recep_flat/"
