+++ test/encodings/onehot/test_oneHotReceptorEncoder.py
@@ -8,6 +8,7 @@ from source.data_model.receptor.receptor_sequence.ReceptorSequence import Recept
 from source.encodings.EncoderParams import EncoderParams
 from source.encodings.onehot.OneHotEncoder import OneHotEncoder
 from source.environment.EnvironmentSettings import EnvironmentSettings
+from source.environment.Label import Label
 from source.environment.LabelConfiguration import LabelConfiguration
 from source.util.PathBuilder import PathBuilder
 
@@ -40,7 +41,8 @@ class TestOneHotSequenceEncoder(TestCase):
         dataset, lc = self._construct_test_dataset(path)
 
         encoder = OneHotEncoder.build_object(dataset, **{"use_positional_info": False,
-                                                         "distance_to_seq_middle": 6})
+                                                         "distance_to_seq_middle": 6,
+                                                         "flatten": False})
 
         encoded_data = encoder.encode(dataset, EncoderParams(
             result_path=f"{path}encoded/",
@@ -64,5 +66,53 @@ class TestOneHotSequenceEncoder(TestCase):
                              [onehot_a, onehot_t, onehot_a, onehot_empty])
         self.assertListEqual([list(item) for item in encoded_data.encoded_data.examples[1, 1]],
                              [onehot_a, onehot_t, onehot_t, onehot_empty])
+
+        shutil.rmtree(path)
+
+    def construct_test_flatten_dataset(self, path):
+        receptors = [TCABReceptor(alpha=ReceptorSequence(amino_acid_sequence="AAATTT", identifier="1a"),
+                                  beta=ReceptorSequence(amino_acid_sequence="ATATAT", identifier="1b"),
+                                  metadata={"l1": 1},
+                                  identifier="1"),
+                     TCABReceptor(alpha=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2a"),
+                                  beta=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2b"),
+                                  metadata={"l1": 2},
+                                  identifier="2"),
+                     TCABReceptor(alpha=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2a"),
+                                  beta=ReceptorSequence(amino_acid_sequence="AAAAAA", identifier="2b"),
+                                  metadata={"l1": 2},
+                                  identifier="2")]
+
+        return ReceptorDataset.build(receptors, 10, "{}receptors.pkl".format(path))
+
+    def test_receptor_flattened(self):
+        path = EnvironmentSettings.root_path + "test/tmp/onehot_recep_flat/"
+
+        PathBuilder.build(path)
+
+        dataset = self.construct_test_flatten_dataset(path)
+
+        encoder = OneHotEncoder.build_object(dataset, **{"use_positional_info": False, "distance_to_seq_middle": None,
+                                                         "flatten": True})
+
+        encoded_data = encoder.encode(dataset, EncoderParams(
+            result_path=path,
+            label_config=LabelConfiguration([Label(name="l1", values=[1, 0], positive_class="1")]),
+            pool_size=1,
+            learn_model=True,
+            model={},
+            filename="dataset.pkl"
+        ))
+
+        self.assertTrue(isinstance(encoded_data, ReceptorDataset))
+
+        onehot_a = [1.0] + [0.0] * 19
+        onehot_t = [0.0] * 16 + [1.0] + [0] * 3
+
+        self.assertListEqual(list(encoded_data.encoded_data.examples[0]), onehot_a+onehot_a+onehot_a+onehot_t+onehot_t+onehot_t+onehot_a+onehot_t+onehot_a+onehot_t+onehot_a+onehot_t)
+        self.assertListEqual(list(encoded_data.encoded_data.examples[1]), onehot_a*12)
+        self.assertListEqual(list(encoded_data.encoded_data.examples[2]), onehot_a*12)
+
+        self.assertListEqual(list(encoded_data.encoded_data.feature_names), [f"{chain}_{pos}_{char}" for chain in ("alpha","beta") for pos in range(6) for char in EnvironmentSettings.get_sequence_alphabet()])
 
         shutil.rmtree(path)
